moved the socket actors to a dedicated actors package
