+++ app/controllers/InstanceRegistryController.scala
@@ -30,7 +30,7 @@ import play.api.mvc._
 import akka.stream.Materializer
 import akka.stream.scaladsl.{Flow, Keep, Sink, Source}
 import play.api.libs.streams.ActorFlow
-
+import actors.{ClientSocketActor, PublishSocketMessageActor}
 
 trait MyExecutionContext extends ExecutionContext
